requirements.txt
appdirs==1.4.0
astroid==1.4.9
coreapi==2.2.4
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.9.10
django-braces==1.11.0
django-filter==1.0.1
django-filters==0.2.1
django-oauth-toolkit==0.11.0
django-rest-swagger==2.1.1
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
gunicorn==19.6.0
isort==4.2.5
itypes==1.1.0
Jinja2==2.9.5
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mccabe==0.6.1
oauthlib==1.1.2
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
PyJWT==1.4.2
pylint==1.6.5
pyparsing==2.1.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
whitenoise==3.3.0
wrapt==1.10.8
